Crate responses

This crate contains all requests and responses used by the frontend of The Bottom Line.

Enums

Connect

The connect response. The very first thing a client should send is this request.

DirectResponse

A response type that a player receives after performing an action. Can either be an error or a confirmation that the action was succesful, including the data needed to update the UI accordingly.

FrontendRequest

Requests that are sent from the frontend, to be handled by the backend.

ResponseError

The general error type that can be sent back in a response.

ResyncData

Custom data used for resyncing a client

RoomCloseReason

Reasons for which a room might have been closed.

UniqueResponse

A response type that is meant for every other player when one player performs an action.