responses

Enum DirectResponse

pub enum DirectResponse {

    Error {
        message: String,
        source: ResponseError,
    },
    YouStartedGame,
    YouSelectedCharacter {
        character: Character,
    },
    YouFiredCharacter {
        character: Character,
    },
    YouTerminateCreditCharacter {
        character: Character,
    },
    YouPaidBanker {
        banker_id: PlayerId,
        new_banker_cash: u8,
        your_new_cash: u8,
        paid_amount: u8,
        sold_assets: Vec<SoldAssetToPayBanker>,
        issued_liabilities: Vec<IssuedLiabilityToPayBanker>,
    },
    YouSelectCardBankerTarget {
        assets: Vec<SoldAssetToPayBanker>,
        liabilities: Vec<IssuedLiabilityToPayBanker>,
    },
    YouRegulatorOptions {
        options: Vec<RegulatorSwapPlayer>,
        character: Character,
        perk: String,
    },
    YouSwapDeck {
        cards_to_draw: usize,
    },
    YouSwapPlayer {
        new_cards: Vec<Either<Asset, Liability>>,
        target_player_id: PlayerId,
    },
    YouAreDivesting {
        options: Vec<DivestPlayer>,
        character: Character,
        perk: String,
    },
    YouDrewCard {
        card: Either<Asset, Liability>,
        can_draw_cards: bool,
        can_give_back_cards: bool,
    },
    YouPutBackCard {
        card_idx: usize,
        can_draw_cards: bool,
        can_give_back_cards: bool,
    },
    YouCharacterAbility {
        character: Character,
        perk: String,
    },
    YouBonusCash {
        cash: u8,
    },
    YouBoughtAsset {
        asset: Asset,
        card_idx: usize,
        market_change: Option<MarketChange>,
    },
    YouIssuedLiability {
        liability: Liability,
        card_idx: usize,
    },
    YouAreFiringSomeone {
        characters: Vec<Character>,
        character: Character,
        perk: String,
    },
    YouDivestedAnAsset {
        target_id: PlayerId,
        asset_idx: usize,
        gold_cost: u8,
    },
    YouAreTerminatingSomeone {
        characters: Vec<Character>,
        character: Character,
        perk: String,
    },
    YouRedeemedLiability {
        liability_idx: usize,
    },
    YouEndedTurn,
    YouJoinedGame {
        username: String,
        channel: String,
    },
    YouRejoined,
    YouResynced {
        id: PlayerId,
        cash: u8,
        hand: Vec<Either<Asset, Liability>>,
        assets: Vec<Asset>,
        liabilities: Vec<Liability>,
        player_info: Vec<PlayerInfo>,
        market: Market,
        phase: ResyncData,
    },
    YouMinusedIntoPlus {
        color: Color,
        new_market: Market,
        new_score: f64,
    },
    YouSilveredIntoGold {
        old_asset_data: Option<SilverIntoGoldData>,
        new_asset_data: Option<SilverIntoGoldData>,
        new_score: f64,
    },
    YouChangedAssetColor {
        old_asset_data: Option<ChangeAssetColorData>,
        new_asset_data: Option<ChangeAssetColorData>,
        new_score: f64,
    },
    YouConfirmedAssetAbility {
        asset_idx: usize,
    },

}

A response type that a player receives after performing an action. Can either be an error or a confirmation that the action was succesful, including the data needed to update the UI accordingly.

Variants

Error

An error returned when the action was not succesful.

Fields

message: String

The error message.

source: ResponseError

The error type.

YouStartedGame

Confirmation that this player started the game.

YouSelectedCharacter

Confirmation that this player selected a character.

Fields

character: Character

The character this player selected.

YouFiredCharacter

Confirmation that this player fired a character.

Fields

character: Character

The character that was fired.

YouTerminateCreditCharacter

Confirmation that this character’s credit line is terminated.

Fields

character: Character

The character who’s credit line was terminated.

YouPaidBanker

Confirmation that you paid some gold to the banker.

Fields

banker_id: PlayerId

The id of the player who is the banker.

new_banker_cash: u8

The new cash balance of the banker.

your_new_cash: u8

The new cash balance of the player that was targeted by the banker.

paid_amount: u8

The amount of gold paid.

sold_assets: Vec<SoldAssetToPayBanker>

A list of assets to be sold to pay off the banker.

issued_liabilities: Vec<IssuedLiabilityToPayBanker>

A list of liabilities to be issued to pay off the banker.

YouSelectCardBankerTarget

Confirmation that you selected someone as target as the banker.

Fields

assets: Vec<SoldAssetToPayBanker>

A list of assets to be sold to pay off the banker.

liabilities: Vec<IssuedLiabilityToPayBanker>

A list of liabilities to be issued to pay off the banker.

YouRegulatorOptions

Confirmation that this player was succesful in getting regulator options

Fields

options: Vec<RegulatorSwapPlayer>

The options this player has to swap with other players.

character: Character

Always Character::Regulator

perk: String

A string containing information about what this character is allowed to do.

YouSwapDeck

Confirmation that this player can proceed swapping with the deck.

Fields

cards_to_draw: usize

The amount of cards this player may draw from the deck.

YouSwapPlayer

Confirmation that this player was succesful in swapping with a player.

Fields

new_cards: Vec<Either<Asset, Liability>>

The new hand of the player.

target_player_id: PlayerId

The id of the player you swapped cards with

YouAreDivesting

Confirmation that this player is now forcing another player to divest.

Fields

options: Vec<DivestPlayer>

A list of cards for each player, which can either be or not be divested.

character: Character

Always Character::Stakeholder

perk: String

A string containing information about what this character is allowed to do.

YouDrewCard

Confirmation that this player drew a card.

Fields

card: Either<Asset, Liability>

The card that was drawn

can_draw_cards: bool

Whether this player can draw another card.

can_give_back_cards: bool

Whether this player should still give back any cards.

YouPutBackCard

Confirmation that this player put back a card.

Fields

card_idx: usize

The index of the card this player put back.

can_draw_cards: bool

Whether this player can draw another card.

can_give_back_cards: bool

Whether this player should still give back any cards.

YouCharacterAbility

Confirmation that this player is trying to use their character ability.

Fields

character: Character

The character of the player.

perk: String

A string containing information about what this player is allowed to do.

YouBonusCash

Confirmation you received your bonus cash and how much.

Fields

cash: u8

The amount of cash received

YouBoughtAsset

Confirmation that this player bought an asset.

Fields

asset: Asset

The asset this player bought.

card_idx: usize

The index of the asset in the player’s hand this player bought.

market_change: Option<MarketChange>

If the market changed, a list of events and a new market is returned.

YouIssuedLiability

Confirmation that this player issued a liability.

Fields

liability: Liability

The liability the player issued.

card_idx: usize

The index of the liability in the player’s hand this player issued.

YouAreFiringSomeone

Confirmation that this player

Fields

characters: Vec<Character>

The list of available characters to fire.

character: Character

Always Character::Shareholder

perk: String

A string containing information on what this character is allowed to do.

YouDivestedAnAsset

Confirmation that this player divested an asset of another player.

Fields

target_id: PlayerId

The id of the player who is forced to divest one of their assets.

asset_idx: usize

The index of the asset they are forced to divest.

gold_cost: u8

The amount of gold it cost to divest this asset.

YouAreTerminatingSomeone

Confirmation that this player is terminating the credit of another player.

Fields

characters: Vec<Character>

A list of characters whose credit can be terminated.

character: Character

Always Character::Banker

perk: String

A string containing information on what this character is allowed to do.

YouRedeemedLiability

Confirmation that this player redeemed a liability.

Fields

liability_idx: usize

The index of the liability that was redeemed.

YouEndedTurn

Confirmation that this player ended their turn.

YouJoinedGame

Confirmation that you joined the lobby.

Fields

username: String

Your connected username.

channel: String

The channel you’re connected to.

YouRejoined

Inform the client that they are rejoining

YouResynced

Deliver data to the client to sync them back to the game state

Fields

id: PlayerId

This player’s personal id.

cash: u8

The amount of cash this player has.

hand: Vec<Either<Asset, Liability>>

The player’s hand.

assets: Vec<Asset>

The assets already played by the player

liabilities: Vec<Liability>

The liabilities already played by the player

player_info: Vec<PlayerInfo>

Public info about every other player.

market: Market

The current market.

phase: ResyncData

A response containing the current gamestate

YouMinusedIntoPlus

Confirms that this player changed one of their market colors.

Fields

color: Color

The market color that was changed,

new_market: Market

The new market for this player.

new_score: f64

The updated player score.

YouSilveredIntoGold

Confirms that this player changed the silver of one of their cards into gold.

Fields

old_asset_data: Option<SilverIntoGoldData>

The data of the old asset that should be updated. When no card was selected before this action, this value will be None.

new_asset_data: Option<SilverIntoGoldData>

The data of the new asset that should be updated. When a card was deselected, this value will be None.

new_score: f64

The updated player score.

YouChangedAssetColor

Confirms that this player changed the color of one of their assets.

Fields

old_asset_data: Option<ChangeAssetColorData>

The data of the old asset that should be updated. When no card was selected before this action, this value will be None.

new_asset_data: Option<ChangeAssetColorData>

The data of the new asset that should be updated. When a card was deselected, this value will be None.

new_score: f64

The updated player score.

YouConfirmedAssetAbility

Confirms that this player confirmed an asset ability’s choice.

Fields

asset_idx: usize

The asset the player confirmed their choice for.