responses

Enum UniqueResponse

pub enum UniqueResponse {

    PlayersInLobby {
        changed_player: String,
        usernames: Vec<String>,
    },
    StartGame {
        id: PlayerId,
        cash: u8,
        hand: Vec<Either<Asset, Liability>>,
        player_info: Vec<PlayerInfo>,
        initial_market: Market,
    },
    SelectingCharacters {
        chairman_id: PlayerId,
        selectable_characters: Option<Vec<Character>>,
        open_characters: Vec<Character>,
        closed_character: Option<Character>,
        turn_order: Vec<PlayerId>,
    },
    SelectedCharacter {
        currently_picking_id: Option<PlayerId>,
        selectable_characters: Option<Vec<Character>>,
        closed_character: Option<Character>,
    },
    TurnStarts {
        player_turn: PlayerId,
        player_turn_cash: u8,
        draws_n_cards: u8,
        gives_back_n_cards: u8,
        playable_assets: PlayableAssets,
        playable_liabilities: u8,
        player_character: Character,
        skipped_characters: Vec<Character>,
    },
    PlayerTargetedByBanker {
        player_turn: PlayerId,
        cash_to_be_paid: u8,
        is_possible_to_pay_banker: bool,
    },
    SelectedCardsBankerTarget {
        assets: Vec<SoldAssetToPayBanker>,
        liability_count: usize,
    },
    DrewCard {
        player_id: PlayerId,
        card_type: CardType,
    },
    PutBackCard {
        player_id: PlayerId,
        card_type: CardType,
    },
    BoughtAsset {
        player_id: PlayerId,
        asset: Asset,
        card_idx: usize,
        market_change: Option<MarketChange>,
    },
    IssuedLiability {
        player_id: PlayerId,
        liability: Liability,
        card_idx: usize,
    },
    RedeemedLiability {
        player_id: PlayerId,
        liability_idx: usize,
    },
    PlayerGotBonusCash {
        player_id: PlayerId,
        cash: u8,
    },
    ShareholderIsFiring {},
    FiredCharacter {
        player_id: PlayerId,
        character: Character,
    },
    TerminatedCreditCharacter {
        player_id: PlayerId,
        character: Character,
    },
    PlayerPaidBanker {
        banker_id: PlayerId,
        player_id: PlayerId,
        new_banker_cash: u8,
        new_target_cash: u8,
        paid_amount: u8,
        sold_assets: Vec<SoldAssetToPayBanker>,
        issued_liabilities: Vec<IssuedLiabilityToPayBanker>,
    },
    RegulatorSwappedYourCards {
        new_cards: Vec<Either<Asset, Liability>>,
    },
    SwappedWithPlayer {
        regulator_id: PlayerId,
        target_id: PlayerId,
    },
    SwappedWithDeck {
        asset_count: usize,
        liability_count: usize,
    },
    AssetDivested {
        player_id: PlayerId,
        target_id: PlayerId,
        asset_idx: usize,
        paid_gold: u8,
    },
    TurnEnded {
        player_id: PlayerId,
    },
    GameEnded {
        scores: Vec<PlayerScore>,
    },
    Rejoined {
        player_id: PlayerId,
    },
    MinusedIntoPlus {
        player_id: PlayerId,
        new_market: Market,
        new_score: f64,
    },
    SilveredIntoGold {
        player_id: PlayerId,
        old_asset_data: Option<SilverIntoGoldData>,
        new_asset_data: Option<SilverIntoGoldData>,
        new_score: f64,
    },
    ChangedAssetColor {
        player_id: PlayerId,
        old_asset_data: Option<ChangeAssetColorData>,
        new_asset_data: Option<ChangeAssetColorData>,
        new_score: f64,
    },
    ConfirmedAssetAbility {
        player_id: PlayerId,
        asset_idx: usize,
    },
    RoomClosed {
        channel: String,
        reason: RoomCloseReason,
    },

}

A response type that is meant for every other player when one player performs an action.

Variants

PlayersInLobby

When someone joins or leaves, this response is sent which includes information on the updated list of players.

Fields

changed_player: String

The name of the player that joined/left.

usernames: Vec<String>

The new list of usernames.

StartGame

Sent when the game was started.

Fields

id: PlayerId

This player’s personal id.

cash: u8

The amount of cash this player gets.

hand: Vec<Either<Asset, Liability>>

The player’s hand.

player_info: Vec<PlayerInfo>

Public info about every other player.

initial_market: Market

The market at the start of the game.

SelectingCharacters

Sent when a SelectingCharacters stage begins.

Fields

chairman_id: PlayerId

The id of the chairman, or the person who selects a character first.

selectable_characters: Option<Vec<Character>>

If it’s this player’s turn, a list of characters that can be selected.

open_characters: Vec<Character>

A list of characters that cannot be selected by anyone.

closed_character: Option<Character>

A character that only the chairman can see, but not select.

turn_order: Vec<PlayerId>

The order each player selects a character in.

SelectedCharacter

Sent when someone selected a character.

Fields

currently_picking_id: Option<PlayerId>

The id of the player that’s currently selecting.

selectable_characters: Option<Vec<Character>>

If it’s this player’s turn, a list of characters that can be selected.

closed_character: Option<Character>

A character that only the chairman can see, but not select.

TurnStarts

Sent when someone’s turn starts.

Fields

player_turn: PlayerId

Id of the player whose turn it is

player_turn_cash: u8

Extra cash received by the player whose turn it is

draws_n_cards: u8

The amount of cards this player draws.

gives_back_n_cards: u8

The amount of cards this player gives back.

playable_assets: PlayableAssets

The amount of assets this player can play, where each color asset has a different ‘unit cost’ attached to it.

playable_liabilities: u8

The amount of liabilities this player can play.

player_character: Character

The character of this player.

skipped_characters: Vec<Character>

A list of characters which were called but were not available.

PlayerTargetedByBanker

Sent when a player is targed by the banker on their turn

Fields

player_turn: PlayerId

Id of the player whose turn it is.

cash_to_be_paid: u8

Amount of cash to be paid to banker.

is_possible_to_pay_banker: bool

Amount of cash to be paid to banker.

SelectedCardsBankerTarget

Sent when a player selects or unselects an asset to sell or liability to issue when paying off the banker.

Fields

assets: Vec<SoldAssetToPayBanker>

A list of assets to be sold to pay off the banker.

liability_count: usize

The number of liabilities the player is set to issue to pay off the banker.

DrewCard

Sent when someone drew a card.

Fields

player_id: PlayerId

The id of the player who drew a card.

card_type: CardType

The type of card this player drew.

PutBackCard

Sent when someone put back a card.

Fields

player_id: PlayerId

The id of the player who put back a card.

card_type: CardType

The type of card this player put back.

BoughtAsset

Sent when a player bought an asset.

Fields

player_id: PlayerId

The id of the player who bought an asset.

asset: Asset

The asset this player bought.

card_idx: usize

The index of the asset in the player’s hand that the player bought.

market_change: Option<MarketChange>

If buying the asset changed the market, sends a list of events as well as the new market.

IssuedLiability

Sent when a player issued a liability.

Fields

player_id: PlayerId

The id of the player who issued a liability

liability: Liability

The liability this player issued.

card_idx: usize

The index of the liability in the player’s hand that the player issued.

RedeemedLiability

Sent when a player

Fields

player_id: PlayerId

The id of the player who

liability_idx: usize

The index of the liability this player redeemed.

PlayerGotBonusCash

Player got their characters bonus gold.

Fields

player_id: PlayerId

PlayerId of the player who got the bonus gold.

cash: u8

Amount of gold the player receiced.

ShareholderIsFiring

Sent when the shareholder is in the process of firing someone.

FiredCharacter

Sent when the shareholder fired a character.

Fields

player_id: PlayerId

The id of the player who fired someone.

character: Character

The character which was fired.

TerminatedCreditCharacter

Sent when a characters credit line has been terminated

Fields

player_id: PlayerId

The id of the player who teminated the credit line someone.

character: Character

The character who’s credit line was terminated.

PlayerPaidBanker

Sent when a player decided on a list of assets and liabilities to pay off the banker.

Fields

banker_id: PlayerId

The id of the player who is the banker this round.

player_id: PlayerId

The id of the player who is the banker.

new_banker_cash: u8

The new cash balance of the banker.

new_target_cash: u8

The new cash balance of the player that was targeted by the banker.

paid_amount: u8

The amount of gold paid.

sold_assets: Vec<SoldAssetToPayBanker>

A list of assets to be sold to pay off the banker.

issued_liabilities: Vec<IssuedLiabilityToPayBanker>

A list of liabilities to be issued to pay off the banker.

RegulatorSwappedYourCards

Sent when the regulator swapped their hand with this player.

Fields

new_cards: Vec<Either<Asset, Liability>>

This player’s new hand.

SwappedWithPlayer

Sent when the regulator swapped their hand with another player.

Fields

regulator_id: PlayerId

The id of the regulator.

target_id: PlayerId

The id of the player the regulator swapped their hands with.

SwappedWithDeck

Sent when the regulator swapped a number of cards with the deck.

Fields

asset_count: usize

The amount of assets the regulator drew from the deck.

liability_count: usize

The amount of liabilities the regulator drew from the deck.

AssetDivested

Sent when the stakeholder forced another player to divest an asset.

Fields

player_id: PlayerId

The id of the stakeholder.

target_id: PlayerId

The id of the player who is forced to divest one of their assets.

asset_idx: usize

The index of the asset they are forced to divest.

paid_gold: u8

The amount of gold the stakeholder paid to divest this asset.

TurnEnded

Sent when someone’s turn ended

Fields

player_id: PlayerId

The id of the player whose turn ended.

GameEnded

Sent when the game ended.

Fields

scores: Vec<PlayerScore>

A list of player scores.

Rejoined

Sent when rejoin request is acknowledged

Fields

player_id: PlayerId

Id of the rejoining player

MinusedIntoPlus

Confirms that a player changed one of their market colors.

Fields

player_id: PlayerId

The id of the player which changed one of their market colors.

new_market: Market

The new market for the player that performed the action,

new_score: f64

The updated player score.

SilveredIntoGold

Confirms that a player changed the silver of one of their cards into gold.

Fields

player_id: PlayerId

The id of the player which changed the silver of one of their cards into gold.

old_asset_data: Option<SilverIntoGoldData>

The data of the old asset that should be updated. When no card was selected before this action, this value will be None.

new_asset_data: Option<SilverIntoGoldData>

The data of the new asset that should be updated. When a card was deselected, this value will be None.

new_score: f64

The updated player score.

ChangedAssetColor

Confirms that a player changed the color of one of their assets.

Fields

player_id: PlayerId

The id of the player which changed the color of one of their assets.

old_asset_data: Option<ChangeAssetColorData>

The data of the old asset that should be updated. When no card was selected before this action, this value will be None.

new_asset_data: Option<ChangeAssetColorData>

The data of the new asset that should be updated. When a card was deselected, this value will be None.

new_score: f64

The updated player score.

ConfirmedAssetAbility

Confirms that a player confirmed an asset ability’s choice.

Fields

player_id: PlayerId

The id of the player which confirmed an asset ability’s choice.

asset_idx: usize

The asset the player confirmed their choice for.

RoomClosed

Used internally to gracefully let everyone know a room closed for any reason.

Fields

channel: String

The channel that was closed.

reason: RoomCloseReason

The reason for which it was closed.